refactor(server): assemble pdv-server core in wire.ts + channel partition + stdio CLI entry
server/wire.ts now owns the entire server-side assembly that lived in
index.ts: session closures (active kernel/project, generation, pending
module imports), per-kernel state maps, the server registrars (including
two new ones: ipc-register-config.ts for the relocated config.get/set and
ipc-register-gui-files.ts for guiEditor.read/save), comm push forwarding
(push-based; cacheKernelSaveResults intact), the autosave idle listener,
MCP server + cell-RPC construction, and mcp:getStatus / cells:respond
served from the invoke registry. index.ts shrinks to the shell wiring
hub: shell registrars, the window-bound push closure (which now owns the
BROADCAST_PUSH_CHANNELS fan-out to child windows), the confirm closure,
the ipcMain mirror, and async server-state access for shell consumers.

ipc.ts gains the SHELL_CHANNELS / SERVER_CHANNELS partition plus push
classifications, enforced by a completeness test; a wire test asserts the
registry after wiring is exactly SERVER_CHANNELS. script.edit moves
shell-side into the launchers registrar (async getConfig/resolveTreeFile
deps); auto-updater reads/writes lastUpdateCheck through the config
channels. server/server-main.ts is the stdio CLI entry (console rebound
to stderr first, env contract, RpcServer serving the registry) — smoke
tested live: hello seq 0, ping, config:get, shutdown, clean exit 0.
sessionReset (for pdv.rpc.sessionReset) restores just-wired state:
closures cleared, kernel working dirs removed, autosave timer stopped.

Behavior deltas: the MCP singleton's push/confirm/hooks are trampolines
into the current wire, fixing a latent macOS window-re-create bug where
MCP kept the first window's stale hooks; mcp.getStatus before server
start resolves {running:false} instead of rejecting; modules-registrar
treeChanged pushes now also reach child windows (fan-out is by channel);
cell-RPC's no-window fast-fail became the request timeout; server-main's
confirm auto-cancels until the reverse-RPC confirm lands with the flip.

Gates: 1041 unit tests green (partition, wire, config, gui-files tests
new); typecheck x3, knip, lint at baseline; Playwright e2e 36 passed /
25 skipped with one kernel-handshake contention flake that passes in
isolation. CLAUDE.md documents the partition rule and new layout.
