+++ electron/main/mcp/cell-rpc.ts
@@ -1,28 +1,33 @@
 /**
- * cell-rpc.ts — Main-side RPC client for the renderer's code-cell state.
+ * cell-rpc.ts — Server-side RPC client for the renderer's code-cell state.
  *
  * Code cells live only in the renderer's React state (ARCHITECTURE.md §15.8).
- * The MCP cell tools therefore use a request/response pattern: main pushes a
- * request to the renderer, the renderer answers via `cells:respond` keyed by
- * the request id. One-way pushes apply writes.
+ * The MCP cell tools therefore use a request/response pattern: the server
+ * pushes a request to the renderer over `IPC.push.cellsRequest`, the
+ * renderer answers via the `cells:respond` invoke channel keyed by the
+ * request id. One-way `IPC.push.cellWrite` pushes apply writes.
  *
  * Responsibilities
- * - Owns the `ipcMain.handle(IPC.cells.respond)` handler — the renderer's
- *   reply lands here and resolves the matching pending promise.
+ * - Correlates `cells:respond` replies (delivered by the wiring layer via
+ *   {@link CellRpcClient.deliver}) to pending requests, with a timeout.
  * - Provides typed `list()` / `read(tabId)` / `write(payload)` methods for
  *   tool code that does not want to think about request ids or timeouts.
  *
  * What it does NOT do
+ * - It does not register the `cells:respond` invoke handler — `server/wire.ts`
+ *   does, forwarding replies into {@link CellRpcClient.deliver}. Keeping
+ *   registration in the wiring layer means re-wiring (macOS window
+ *   re-creation) re-registers it alongside every other server channel.
  * - It is not an MCP tool itself; it is a helper consumed by the cell tools
- *   in `tools/execution.ts` (Phase 2).
- * - It does not interpret cell content beyond the renderer-shape types.
+ *   in `tools/execution.ts`.
+ * - It does not know whether a renderer window exists: pushes go through the
+ *   injected {@link PushSender}, and an unanswerable request fails by
+ *   timeout rather than fast-failing on a missing window.
  *
  * See Also
  * --------
  * ARCHITECTURE.md §15.8 — Renderer Interaction
  */
-
-import { BrowserWindow, ipcMain } from "electron";
 
 import { randomUUID } from "node:crypto";
 
@@ -33,6 +38,7 @@ import {
   type CellsResponse,
   IPC,
 } from "../ipc";
+import type { PushSender } from "../server/invoke-registry";
 
 /** How long a cell request waits before failing if the renderer never replies. */
 const DEFAULT_TIMEOUT_MS = 10_000;
@@ -44,49 +50,32 @@ interface PendingRequest {
 }
 
 /**
- * Bridges main-process callers (the MCP cell tools) to renderer code-cell state.
+ * Bridges server-side callers (the MCP cell tools) to renderer code-cell
+ * state over the injected push sender.
  *
- * The client is window-bound — it sends pushes via the supplied BrowserWindow
- * accessor and matches responses back to pending requests by `requestId`. Call
- * {@link start} once at app boot to register the IPC handler; call {@link stop}
- * during shutdown so the handler is detached cleanly.
+ * Call {@link stop} during shutdown so in-flight requests reject promptly
+ * instead of dangling until their timeouts.
  */
 export class CellRpcClient {
-  private readonly getWindow: () => BrowserWindow | null;
+  private readonly push: PushSender;
   private readonly pending = new Map<string, PendingRequest>();
-  private started = false;
 
   /**
-   * @param getWindow - Accessor for the renderer window that owns the cell
-   *   tabs. Returns `null` when no window is open — calls in that state are
-   *   rejected rather than throwing synchronously.
+   * @param push - Renderer-push sender. In single-process mode this is the
+   *   shell's window-bound closure; in the extracted pdv-server it is the
+   *   transport's seq-stamping sender.
    */
-  constructor(getWindow: () => BrowserWindow | null) {
-    this.getWindow = getWindow;
+  constructor(push: PushSender) {
+    this.push = push;
   }
 
   /**
-   * Register the `cells:respond` IPC handler. Idempotent.
-   *
-   * @returns Nothing.
-   */
-  start(): void {
-    if (this.started) return;
-    ipcMain.handle(IPC.cells.respond, (_event, response: CellsResponse) => {
-      this.deliver(response);
-    });
-    this.started = true;
-  }
-
-  /**
-   * Detach the IPC handler and reject any in-flight requests.
+   * Reject any in-flight requests. Idempotent; the client remains usable
+   * (a later request simply starts a new pending entry).
    *
    * @returns Nothing.
    */
   stop(): void {
-    if (!this.started) return;
-    ipcMain.removeHandler(IPC.cells.respond);
-    this.started = false;
     for (const pending of this.pending.values()) {
       clearTimeout(pending.timer);
       pending.reject(new Error("Cell RPC client stopped"));
@@ -98,7 +87,7 @@ export class CellRpcClient {
    * Ask the renderer for the current list of cell tabs.
    *
    * @returns Tabs + active-tab metadata.
-   * @throws {Error} When the renderer is unreachable or does not reply.
+   * @throws {Error} When the renderer does not reply in time.
    */
   async list(): Promise<CellListResult> {
     return (await this.request({ op: "list" })) as CellListResult;
@@ -120,23 +109,34 @@ export class CellRpcClient {
    *
    * @param payload - The write payload.
    * @returns Nothing.
-   * @throws {Error} When no renderer window is available.
    */
   write(payload: CellWritePush): void {
-    const win = this.getWindow();
-    if (!win) {
-      throw new Error("Cell write failed: no renderer window");
+    this.push(IPC.push.cellWrite, payload);
+  }
+
+  /**
+   * Resolve a pending request from a renderer `cells:respond` reply.
+   * Called by the wiring layer's invoke handler; late replies after a
+   * timeout are dropped silently.
+   *
+   * @param response - The renderer's reply payload.
+   * @returns Nothing.
+   */
+  deliver(response: CellsResponse): void {
+    const pending = this.pending.get(response.requestId);
+    if (!pending) return; // late reply after timeout — drop silently
+    this.pending.delete(response.requestId);
+    clearTimeout(pending.timer);
+    if (response.ok && response.result) {
+      pending.resolve(response.result);
+    } else {
+      pending.reject(new Error(response.error ?? "Cell request failed"));
     }
-    win.webContents.send(IPC.push.cellWrite, payload);
   }
 
   private async request(
     body: { op: "list" | "read"; tabId?: number },
   ): Promise<CellListResult | CellReadResult> {
-    const win = this.getWindow();
-    if (!win) {
-      throw new Error(`Cell ${body.op} failed: no renderer window`);
-    }
     return new Promise<CellListResult | CellReadResult>((resolve, reject) => {
       const requestId = randomUUID();
       const timer = setTimeout(() => {
@@ -149,19 +149,7 @@ export class CellRpcClient {
         );
       }, DEFAULT_TIMEOUT_MS);
       this.pending.set(requestId, { resolve, reject, timer });
-      win.webContents.send(IPC.push.cellsRequest, { requestId, ...body });
+      this.push(IPC.push.cellsRequest, { requestId, ...body });
     });
-  }
-
-  private deliver(response: CellsResponse): void {
-    const pending = this.pending.get(response.requestId);
-    if (!pending) return; // late reply after timeout — drop silently
-    this.pending.delete(response.requestId);
-    clearTimeout(pending.timer);
-    if (response.ok && response.result) {
-      pending.resolve(response.result);
-    } else {
-      pending.reject(new Error(response.error ?? "Cell request failed"));
-    }
   }
 }
